Enum gic::ArmGic

pub enum ArmGic {
    V2(ArmGicV2),
    V3(ArmGicV3),
}

Arm Generic Interrupt Controller

The GIC is an extension to ARMv8 which allows routing and filtering interrupts in a single or multi-core system.

Variants

V2(ArmGicV2)

V3(ArmGicV3)

Implementations

impl ArmGic

pub fn get_spi_target(&self, int: InterruptNumber) -> TargetCpu

The GIC can be configured to route Shared-Peripheral Interrupts (SPI) either to a specific CPU or to any PE that is ready to process them, i.e. not handling another higher-priority interrupt.

pub fn set_spi_target(&mut self, int: InterruptNumber, target: TargetCpu)

The GIC can be configured to route Shared-Peripheral Interrupts (SPI) either to a specific CPU or to any PE that is ready to process them, i.e. not handling another higher-priority interrupt.

impl ArmGic

pub fn init(
page_table: &mut PageTable,
version: Version
) -> Result<Self, &'static str>

pub fn send_ipi(&mut self, int_num: InterruptNumber, target: TargetCpu)

Sends an inter processor interrupt (IPI), also called software generated interrupt (SGI).

note: on Aarch64, IPIs must have a number below 16 on ARMv8

pub fn acknowledge_interrupt(&mut self) -> (InterruptNumber, Priority)

Acknowledge the currently serviced interrupt and fetches its number

pub fn end_of_interrupt(&mut self, int: InterruptNumber)

Performs priority drop for the specified interrupt

pub fn get_interrupt_state(&self, int: InterruptNumber) -> Enabled

Will that interrupt be forwarded by the distributor?

pub fn set_interrupt_state(&mut self, int: InterruptNumber, enabled: Enabled)

Enables or disables the forwarding of a particular interrupt in the distributor

pub fn get_minimum_priority(&self) -> Priority

Interrupts have a priority; if their priority is lower or equal to this one, they’re queued until this CPU or another one is ready to handle them

pub fn set_minimum_priority(&mut self, priority: Priority)

Interrupts have a priority; if their priority is lower or equal to this one, they’re queued until this CPU or another one is ready to handle them